+++ packages/login-panel/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.3](https://github.com/OBSAU/rtm-kit/compare/@rtm-ui/login-panel@1.0.2...@rtm-ui/login-panel@1.0.3) (2019-01-10)
+
+**Note:** Version bump only for package @rtm-ui/login-panel
+
+
+
+
+
+## [1.0.2](https://github.com/OBSAU/rtm-kit/compare/@rtm-ui/login-panel@1.0.1...@rtm-ui/login-panel@1.0.2) (2019-01-10)
+
+
+### Bug Fixes
+
+* name modules.js to module.js on build ([8a9848c](https://github.com/OBSAU/rtm-kit/commit/8a9848c))
+
+
+
+
+
 ## 1.0.1 (2019-01-09)